// src/external/api/models/LoginResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { JwtTokenResponse } from './JwtTokenResponse';
export type LoginResponse = {
    sessionToken?: JwtTokenResponse;
    refreshToken?: JwtTokenResponse;
};